Create Downloadparticularfikefromzippedfolder.cs
In this modified code, we specify the path to the file we want to download within the zipped folder using the fileToDownload variable. We use the GetEntry method of the ZipArchive class to get the ZipArchiveEntry object representing the file within the zipped folder, and then extract it in the same way as before. If the specified file does not exist within the zipped folder, we throw an exception.