Create Zipfolderintostream
This code uses the Renci.SshNet library to connect to an SFTP server, downloads a zip file located at zipFilePath into memory, extracts the files from the zip archive, and stores the extracted file streams in a list called fileStreams.

Note that the MemoryStream class is used to create an in-memory stream, which can be used to read and write data as if it were a regular stream. The ZipArchive class is used to open the zip archive, and the archive.Entries property is used to get a list of files in the archive. Finally, the entry.Open() method is used to read the contents of each file in the archive into a file stream, which is then added to the fileStreams list.

After processing the file streams, the SFTP connection is closed using the sftp.Disconnect() method.