Create Multi-threading
In this example, we are using the Session class from the WinSCP .NET library to connect to the SFTP server and download files. We are also using the Parallel.ForEach method to download multiple files in parallel using multiple threads.

The GetFiles method is used to download a remote file to a stream. We are using a MemoryStream as the output stream to download the file, and then copying the contents of the MemoryStream to a local file using a FileStream.

Note that this example does not use SSH key authentication, but instead uses a username and password to authenticate with the SFTP server. If you have the option to use SSH key authentication, it is generally considered more secure.